Add mutex for operations on gitRepository
